Silence gcc warnings (signed vs. unsigned, implicit declarations, unused vars).

+++ clients/tools/ldapsearch.c
@@ -729,7 +729,7 @@ getNextPage:
 		i = 0;
 		moreEntries = getchar();
 		while ( moreEntries != EOF && moreEntries != '\n' ) { 
-			if ( i < sizeof(buf) - 1 ) {
+			if ( i < (int)sizeof(buf) - 1 ) {
 				buf[i] = moreEntries;
 				i++;
 			}
